% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECLIPSE.R
\name{ECLIPSE}
\alias{ECLIPSE}
\title{Analyzes scTCR-seq data and returns results in Seurat object. Specifically,
chain pairings across the sample are analyzed to predict chains lost to
sequencing dropout and to decipher whether cells with >2 total TCR chains
are caused by technical artifacts or true biology.}
\usage{
ECLIPSE(
  contig_type = "all",
  contig_folder_paths = NULL,
  contig_file_paths = NULL,
  seurat_object,
  batch,
  donor,
  group,
  original_barcode,
  format = "Blank",
  detailedEclipseStats = FALSE,
  combineTcrArgs = list(FALSE, FALSE, FALSE, TRUE),
  allowDoubleChains = "both",
  returnFilteredContigs = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{contig_type}{tells whether contig_folder_paths or contig_file_paths
should be used to read in contig files and how. To provide file paths,
this should be "manual". To provide folders, this should be "all" or
"filtered" to search for all or filtered contig annotations files.}

\item{contig_folder_paths}{vector of paths to directories that contain TCR
files. If you use this argument, provide either "all" or "filtered" to
contig_type. The order should match the numbers provided in batch.}

\item{contig_file_paths}{vector of TCR file paths. The order should match
the numbers provided in batch. The files can have any name, but they must
be all_contig_annotations.csv or
filtered_contig_annotations.csv format.}

\item{seurat_object}{Seurat object that contains scRNA-seq data.}

\item{batch}{name of column in Seurat meta.data. The column should contain
integers, not characters, that tell which position in the
contig_folder/file_path vectors contains the TCR contigs for the cell.}

\item{donor}{name of column in Seurat meta.data. The column should contain
the donor, regardless of treatment/timepoint/condition.}

\item{group}{name of column in Seurat meta.data. This column should specify
how the final statistics for each clone should be calculated. For example,
should they be calculated per sample, per donor, or among all cells?
This can and often will be the same column as donor.}

\item{original_barcode}{name of column in Seurat meta.data. This must
contain the original 10X barcode for the cell. Typically this will end with
-1, but users should check their TCR contig files to be sure.}

\item{format}{default is blank. This tells how missing data is stored in
the TCR contig files. Current files likely contain "" so "blank" is the
default, but if older contig files are used where any missing data is stored
as "None", then "None" should be provided here.}

\item{detailedEclipseStats}{default is FALSE. Provides detailed stats
relating to EM assignment in Seurat meta.data for each cell if TRUE.}

\item{combineTcrArgs}{arguments that are provided to scRepertoire
combineTCR. This should be a list of 4 logical elements,
with the elements corresponding to the combineTCR arguments.
1 is removeNA, 2 is removeMulti, 3 is filterMulti, 4 is filterNonproductive.
If no clonotype can be predicted with high confidence, then combineTCR
using these arguments will determine the final clonotype. The default is
the combineTCR default where only filterNonproductive is TRUE.}

\item{allowDoubleChains}{specifies whether 2 alpha chains, 2 beta chains,
or either 2 alpha or 2 beta chains are allowed. Default is both.}

\item{returnFilteredContigs}{specifies if filtered contigs should be returned
instead of running the whole ECLIPSE pipeline. Default is FALSE.}

\item{verbose}{specifies whether to print an important
QC graph, a small table, and also notes through the pipeline. Default is
TRUE.}
}
\value{
A Seurat object with TCR data provided in meta.data.
}
\description{
Analyzes scTCR-seq data and returns results in Seurat object. Specifically,
chain pairings across the sample are analyzed to predict chains lost to
sequencing dropout and to decipher whether cells with >2 total TCR chains
are caused by technical artifacts or true biology.
}
\examples{
data(test_seurat)

files <- system.file("extdata/examples", package = "ECLIPSE")
tcr_paths <- list.files(
                files,
                pattern = "\\\\.csv.gz$",
                full.names = TRUE)

cd8_tcr <- ECLIPSE(contig_file_paths = tcr_paths,
                    contig_type = "manual",
                    seurat_object = test_seurat,
                    batch = "index",
                    donor = "donor",
                    group = "donor",
                    original_barcode = "Cell_Barcode")

}
