% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findThreeChainClones.R
\name{findThreeChainClones}
\alias{findThreeChainClones}
\title{Modifies and returns cellranger TCR contig files if clones
contain two alpha or beta chains.}
\usage{
findThreeChainClones(contigs, chain, mode = "standard", printStats = TRUE)
}
\arguments{
\item{contigs}{tibble of contigs from cellranger all or
filtered contig files.}

\item{chain}{"alpha" or "beta". This tells it whether clones with two of
which chain should be looked for.}

\item{mode}{"standard" is the default, where modified contigs are returned.
If "show" is provided, two tables are returned. The first
shows the final set of clones with a double chain of one type, and the
second shows how/why some of those clones were called as such. The second
is pre-filtering, so some clones may not have been called as having 3 total
chains in the end.}

\item{printStats}{specifies whether notes/stats
should be printed. Default is TRUE.}
}
\value{
a tibble of modified contigs,
except if mode show is selected.
}
\description{
Modifies and returns cellranger TCR contig files if clones
contain two alpha or beta chains.
}
\examples{
data(post_qc_contigs)

modified_contigs <- findThreeChainClones(
                        contigs = post_qc_contigs,
                        chain = "a")

}
