% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_seurat}
\alias{test_seurat}
\title{Example minimal Seurat object without TCR information.}
\format{
A Seurat object.
}
\source{
Example data
}
\usage{
test_seurat
}
\description{
Example minimal Seurat object without TCR information.
}
\keyword{datasets}
