% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_object}
\alias{vignette_object}
\title{Example small Seurat object without TCR information.}
\format{
A Seurat object.
}
\source{
Example data
}
\usage{
data(vignette_object)
}
\description{
Example small Seurat object without TCR information.
}
\keyword{datasets}
