% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{post_qc_contigs}
\alias{post_qc_contigs}
\title{Example Post-QC TCR contigs that would be generated within ECLIPSE
and passed to findThreeChainClones.}
\format{
A data frame in the all_contig_annotations.csv format
that is produced by cellranger vdj.
}
\source{
Example data
}
\usage{
post_qc_contigs
}
\description{
Example Post-QC TCR contigs that would be generated within ECLIPSE
and passed to findThreeChainClones.
}
\keyword{datasets}
