% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcrDoubletDetect.R
\name{tcrDoubletDetect}
\alias{tcrDoubletDetect}
\title{Finds and removes doublets from Seurat objects}
\usage{
tcrDoubletDetect(
  object,
  filter = TRUE,
  singleChainLimit = 2,
  totalChainLimit = -1,
  showPlots = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{object}{Seurat object.}

\item{filter}{logical statement whether cells should be filtered or not.
Default is TRUE.}

\item{singleChainLimit}{number of alpha or beta chains that is allowed
before filtering occurs. Must be a integer, with a default of 2.}

\item{totalChainLimit}{number of total TCR chains that is allowed
before filtering occurs. Must be an integer. Default is no threshold.}

\item{showPlots}{logical statement that determines whether plots
are provided that help set additional filtering thresholds based on
RNA counts and TCR chain count. Default is TRUE.}

\item{verbose}{specifies whether notes should be printed. Default is TRUE.}
}
\value{
a Seurat object that is filtered if desired.
}
\description{
Filters out cells with excessive numbers of TCR chains and
helps set thresholds for RNA counts based on TCR chain count.
}
\examples{
data(test_seurat_with_tcr)

cd8_tcr_filtered <- tcrDoubletDetect(
                            test_seurat_with_tcr,
                            singleChainLimit = 2)

}
