% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_seurat_with_tcr}
\alias{test_seurat_with_tcr}
\title{Example Seurat object with TCR information generated by running
the ECLIPSE function.}
\format{
A Seurat object.
}
\source{
Example data
}
\usage{
data(test_seurat_with_tcr)
}
\description{
Example Seurat object with TCR information generated by running
the ECLIPSE function.
}
\keyword{datasets}
